Move function count_py_files to file_helpers.py
