+++ fiasko_bro/file_helpers.py
@@ -1,0 +1,10 @@
+import os
+
+
+def count_py_files(directory):
+    counter = 0
+    for _, _, filenames in os.walk(directory):
+        for name in filenames:
+            if name.endswith('.py'):
+                counter += 1
+    return counter
